+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - ViewPoint:
     - Migrated from dev to dev_vue.
 - Migrator: supports ids with suffix.
+- InputText:
+    - Added change event handler property and disabled property.
 
 ### Changed
 - Eslint: no-undef eslint rule is switched on.